Runtime delete2 (#71)
* stash to review code

* RealTimeDelete

* gona re-merge main

* warning changed to info

* stash

* also updated scenes

* fix assim load bug

+++ GTXDankEngine/Source/Systems/ScriptSystem/ScriptSystem.h
@@ -27,7 +27,7 @@ struct Event
 	{
 		if (affectEverything)
 		{
-			for (int i = 1; i < GameLogicCategories::MAX_CATEGORIES; ++i)
+			for (int i = 1; i < static_cast<int>(GameLogicCategories::MAX_CATEGORIES); ++i)
 			{
 				thingsToEffect.insert(static_cast<GameLogicCategories>(i));
 			}
